refactor : product-slide 이름 수정 및 코드
- product-slide.js 이름 변경
